day 09

+++ src/com/dehnes/adventofcode_2019/day05/Day05.kt
@@ -3,7 +3,7 @@ package com.dehnes.adventofcode_2019.day05
 import com.dehnes.adventofcode_2019.inputText
 import com.dehnes.adventofcode_2019.intcodeComputer
 
-val originalCode = inputText(5).split(",").map { it.toInt() }
+val originalCode = inputText(5).split(",").map { it.toLong() }
 
 fun main() {
     intcodeComputer(originalCode, { 1 }, ::println) // 5074395
